READ DESCRIPTION! Garbage discoveries applied.
- IMPORTANT!: Debugging variable added to each agent (at the beginning
of each class) in order to show/not show their console messages (all set
to false now).
- Added garbage discoveries feature (from Scouts).
- Added COMPLETE communication from Scouts to Central and from Central
to Harv.Coord. in order to detect ONLY the NEW garbage spots.

NOTE: the new garbages are added into the variable "auctionGarbages"
from the Harv.Coord. in order to be ready for auction.