Collision system implemented READ DESCRIPTION
Here are the possible situations:
Scout - Harvester -> Always pass the Harvester
Scout - Scout -> Passes the one with greater ID
Harv - Harv -> Pass the one which has more units of garbage. If it its
the same, pass the one with greater ID

IMPORTANT: This is not optimal at all, so it could be that a Scout moves
in order to avoid a Harvester and the Harvester is not gonna move or is
gonna move in an opposite directon so that the Scout didn't have to be
moved before. But For the moment, it works!